Added `ser/structs` module
