+++ crates/bevy_reflect/src/serde/ser/mod.rs
@@ -3,14 +3,14 @@ pub use serializer::*;
 
 mod serializable;
 mod serializer;
+mod structs;
 
 use crate::{
-    Array, Enum, List, Map, Set, Struct, Tuple, TupleStruct, TypeInfo, TypeRegistry, VariantInfo,
+    Array, Enum, List, Map, Set, Tuple, TupleStruct, TypeInfo, TypeRegistry, VariantInfo,
     VariantType,
 };
 use serde::ser::{
-    Error, SerializeStruct, SerializeStructVariant, SerializeTuple, SerializeTupleStruct,
-    SerializeTupleVariant,
+    Error, SerializeStructVariant, SerializeTuple, SerializeTupleStruct, SerializeTupleVariant,
 };
 use serde::{
     ser::{SerializeMap, SerializeSeq},
@@ -18,59 +18,6 @@ use serde::{
 };
 
 use super::SerializationData;
-
-pub struct StructSerializer<'a> {
-    pub struct_value: &'a dyn Struct,
-    pub registry: &'a TypeRegistry,
-}
-
-impl<'a> Serialize for StructSerializer<'a> {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        let type_info = self
-            .struct_value
-            .get_represented_type_info()
-            .ok_or_else(|| {
-                Error::custom(format_args!(
-                    "cannot get type info for {}",
-                    self.struct_value.reflect_type_path()
-                ))
-            })?;
-
-        let struct_info = match type_info {
-            TypeInfo::Struct(struct_info) => struct_info,
-            info => {
-                return Err(Error::custom(format_args!(
-                    "expected struct type but received {info:?}"
-                )));
-            }
-        };
-
-        let serialization_data = self
-            .registry
-            .get(type_info.type_id())
-            .and_then(|registration| registration.data::<SerializationData>());
-        let ignored_len = serialization_data.map(SerializationData::len).unwrap_or(0);
-        let mut state = serializer.serialize_struct(
-            struct_info.type_path_table().ident().unwrap(),
-            self.struct_value.field_len() - ignored_len,
-        )?;
-
-        for (index, value) in self.struct_value.iter_fields().enumerate() {
-            if serialization_data
-                .map(|data| data.is_field_skipped(index))
-                .unwrap_or(false)
-            {
-                continue;
-            }
-            let key = struct_info.field_at(index).unwrap().name();
-            state.serialize_field(key, &TypedReflectSerializer::new(value, self.registry))?;
-        }
-        state.end()
-    }
-}
 
 pub struct TupleStructSerializer<'a> {
     pub tuple_struct: &'a dyn TupleStruct,
